// source/common/enums/EServiceCategory.ts
export enum EServiceCategory {
  BED_CLEANING = "BED_CLEANING",
  BLANKET_CLEANING = "BLANKET_CLEANING",
  CARPET_CLEANING = "CARPET_CLEANING",
  CHAIR_CLEANING = "CHAIR_CLEANING",
  CURTAIN_CLEANING = "CURTAIN_CLEANING",
  QUILT_CLEANING = "QUILT_CLEANING",
  SOFA_CLEANING = "SOFA_CLEANING",
  VEHICLE_CLEANING = "VEHICLE_CLEANING",
}